gl: Fix incomplete stencil view of depth-stencil texture
- Samplers must use point sampling for stencil views
