+++ rpcs3/GLGSRender.vcxproj
@@ -68,6 +68,7 @@
     <ClInclude Include="Emu\RSX\GL\glutils\pixel_settings.hpp" />
     <ClInclude Include="Emu\RSX\GL\glutils\program.h" />
     <ClInclude Include="Emu\RSX\GL\glutils\ring_buffer.h" />
+    <ClInclude Include="Emu\RSX\GL\glutils\sampler.h" />
     <ClInclude Include="Emu\RSX\GL\glutils\state_tracker.hpp" />
     <ClInclude Include="Emu\RSX\GL\glutils\image.h" />
     <ClInclude Include="Emu\RSX\GL\glutils\sync.hpp" />
@@ -95,6 +96,7 @@
     <ClCompile Include="Emu\RSX\GL\glutils\program.cpp" />
     <ClCompile Include="Emu\RSX\GL\glutils\ring_buffer.cpp" />
     <ClCompile Include="Emu\RSX\GL\glutils\image.cpp" />
+    <ClCompile Include="Emu\RSX\GL\glutils\sampler.cpp" />
     <ClCompile Include="Emu\RSX\GL\GLVertexProgram.cpp" />
     <ClCompile Include="Emu\RSX\GL\GLHelpers.cpp" />
     <ClCompile Include="Emu\RSX\GL\GLPresent.cpp" />
