+++ rpcs3/GLGSRender.vcxproj.filters
@@ -38,6 +38,9 @@
     <ClCompile Include="Emu\RSX\GL\glutils\program.cpp">
       <Filter>glutils</Filter>
     </ClCompile>
+    <ClCompile Include="Emu\RSX\GL\glutils\sampler.cpp">
+      <Filter>glutils</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Emu\RSX\GL\GLTexture.h" />
@@ -90,6 +93,9 @@
       <Filter>glutils</Filter>
     </ClInclude>
     <ClInclude Include="Emu\RSX\GL\glutils\sync.hpp">
+      <Filter>glutils</Filter>
+    </ClInclude>
+    <ClInclude Include="Emu\RSX\GL\glutils\sampler.h">
       <Filter>glutils</Filter>
     </ClInclude>
   </ItemGroup>
